feat(dapp): do not fetch svg, use static img
