+++ dapp/src/app/(protected)/my-names/components/PanelTile.tsx
@@ -13,7 +13,7 @@ import {
 import { normalizeIotaName } from '@iota/iota-names-sdk';
 import clsx from 'clsx';
 
-import { AvatarDisplay } from '@/components/name-record/AvatarDisplay';
+import { NameCardDisplay } from '@/components/name-record/NameCardDisplay';
 
 import { PanelTileType } from './enums';
 
@@ -57,7 +57,7 @@ export function PanelTile({
                                 type={ImageType.BgTransparent}
                                 shape={ImageShape.SquareRounded}
                             >
-                                <AvatarDisplay
+                                <NameCardDisplay
                                     name={name}
                                     size="full"
                                     fallbackUrl="/subname-card-fallback.png"
